refactor: extract extractErrorMessage + consolidate carFilterFn (#249 #250)
- refactor(#249): add extractErrorMessage() to client/src/utils/handlerUtils.js;
  remove the identical inline definition from 5 hook files (useUserActions,
  useCarActions, useMessageActions, useServiceActions, useAccountActions)
  and replace with a single shared import
- refactor(#250): create client/src/utils/filterUtils.js with the canonical
  carFilterFn (includes owner.username for admin search); update both callers
  (useCarsTableData, useAccountTableData) to import from the shared util;
  remove the duplicated definition from accountValidation.js (was missing
  owner.username); remove from carValidation.js and re-export from filterUtils

Closes #249, closes #250
