+++ client/src/pages/account/hooks/useAccountActions.js
@@ -1,6 +1,5 @@
 import { createReqService } from "../../../api/services/messageApi";
-
-const extractErrorMessage = (err) => err?.response?.data?.message ?? err?.message ?? String(err);
+import { extractErrorMessage } from "../../../utils/handlerUtils";
 
 /**
  * Custom hook for account service operations
